move dataviews to generators logic into dataview-client package
