+++ packages/api-types/package.json
@@ -39,5 +39,8 @@
     "rollup-plugin-multi-input": "^1.1.1",
     "rollup-plugin-postcss": "^3.1.8",
     "rollup-plugin-terser": "^7.0.2"
+  },
+  "peerDependencies": {
+    "@globalfishingwatch/data-transforms": "1.1.1"
   }
 }
