+++ packages/dataviews-client/package.json
@@ -41,6 +41,7 @@
     "@types/lodash": "^4.14.168",
     "@types/qs": "^6.9.3",
     "cross-env": "^7.0.3",
+    "d3-scale": "3.2.3",
     "rollup": "^2.33.1",
     "rollup-plugin-multi-input": "^1.1.1",
     "rollup-plugin-postcss": "^3.1.8",
