+++ packages/react-hooks/src/index.ts
@@ -1,8 +1,4 @@
-export {
-  default as useDataviewsGeneratorConfigs,
-  getDataviewsGeneratorConfigs,
-  getGeneratorConfig,
-} from './use-dataviews-layers'
+export { default as useDataviewsGeneratorConfigs } from './use-dataviews-layers'
 export { default as useDebounce } from './use-debounce'
 export { default as useLayerComposer } from './use-layer-composer'
 export { default as useLogin } from './use-login'
@@ -15,4 +11,3 @@ export type {
 export { default as useMapLegend } from './use-map-legend'
 export { default as useTilesState } from './use-tiles-state'
 export { default as useTilesLoading } from './use-tiles-loading'
-export { default as useWorkspace } from './use-workspace'
