--- packages/react-hooks/src/use-workspace/index.ts
@@ -1,1 +1,0 @@
-export { default } from './use-workspace'
